==> Notebooks 2025/3_data_manip_technical_info.ipynb <==
# Data manipulation file after the first scraping step to expand the technical information column

## 1. Library and data import

import pandas as pd
import ast
from datetime import datetime
import re

# Loading the scraped data
df = pd.read_csv('final_scrape_2025.csv')
pd.set_option('display.max_columns', None)

# Displaying the data
df

## 2. Various manipulations and verifications

def safe_eval(val):
    """Convert string to dict if possible, else return empty dict."""
    if val == 'N/A':
        return {}
    try:
        return ast.literal_eval(val)
    except Exception:
        return {}

def is_legislative_proposal(docs):
    if isinstance(docs, str):
        try:
            # Converts string in a list of dictionnaries
            docs = ast.literal_eval(docs)
        except Exception:
            return False

    if isinstance(docs, list) and len(docs) > 0:
        first_doc = docs[0]
        if isinstance(first_doc, dict) and 'Document type' in first_doc:
            return first_doc['Document type'] == 'Legislative proposal'
    return False

df['is_legislative_proposal'] = df['documentation_gateway'].apply(is_legislative_proposal)

# Convert the technical information columns into a dict
df['technical_info'] = df['technical_info'].apply(safe_eval)

# Expand the technical info into separate columns for better accessibility and readability of the variables:
technical_df = pd.json_normalize(df['technical_info'])

# Combine with the original df
df = pd.concat([df.drop(columns=['technical_info']), technical_df], axis=1)

df

# Save the manipulated data
df.to_csv('final_sample_cod_manip_2025.csv')
