==> Notebooks 2025/5_lp_scraped_verifier.ipynb <==
# Identifying which legislative texts have been scraped

## 1. Library and data import

import pandas as pd
import re

df = pd.read_csv('final_sample_cod_legislative_scrape_2025.csv').drop(columns=['Unnamed: 0.1', 'Unnamed: 0'])

df

## 2. Data manipulation

# Some legislative proposals may not have beeen scraped because the website's format has changed since 2014
# Here is a code to check if the legislative proposals begin with "chooose the experimental features you want"
# or if it is an empty value: if so, it means the legislative proposal hasn't been scraped

df['lp_scraped'] = 1  # default value, meaning the legislative proposal has been scraped

for row, data in df.iterrows():
    txt = str(data['legislative_proposal_text']).lower()
    if re.match(r'choose the experimental features you want|nan', txt):
        df.at[row, 'lp_scraped'] = 0 # the legislative proposal has not been scraped

# Checking which rows have ahd their legislative proposals actually scraped
df = df[df['lp_scraped'] == 1]
df = df.reset_index().drop(columns='index')
df
