==> Notebooks 2025/8_splitting_datasets.ipynb <==
# Data manipulation notebook to create the datasets for completed and ongoing procedures

## 1. Library and data import

import pandas as pd

df = pd.read_csv('final_sample_cod_TA_final_act_2025.csv')

## 2. Data manipulation

# Splitting the dataset into four different ones: one for the completed procedures, one for the rejected procedures, 
# one for the lapsed or withdrawn procedures and one for the ongoing procedures

finished_stages = {'Procedure completed', 'Procedure rejected', 'Procedure lapsed or withdrawn'}

df_completed = df[df['Stage reached in procedure'] == 'Procedure completed']
df_rejected = df[df['Stage reached in procedure'] == 'Procedure rejected']
df_lapsed = df[df['Stage reached in procedure'] == 'Procedure lapsed or withdrawn']
df_ongoing = df[~df['Stage reached in procedure'].isin(finished_stages)]

# Saving the data
df_completed.to_csv('sample_cod_completed_2025.csv')
df_rejected.to_csv('sample_cod_rejected_2025.csv')
df_lapsed.to_csv('sample_cod_lapsed_2025.csv')
df_ongoing.to_csv('sample_cod_ongoing_2025.csv')
